enable glance-api to use registry outside of K8s
When running in K8s glance-api and glance-registy containers are
deployed in the same pod and can therefore communicate via the
standard registry port of 9191 on the localhost interface.
However when deploying outside of K8s these two containers can't
communicate over localhost so the glance-registry container needs
to EXPOSE its port and then the glance-api needs to be configured
to point at the ip of the registry container.